requirements/requirements.txt
setuptools
six
keyring==25.3.0
Django==4.2.14
psycopg==3.2.1
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.2.2
djangorestframework==3.15.2
python-dotenv==1.0.1
drf_spectacular==0.27.2
redis==5.0.8
django-redis==5.4.0
celery==5.4.0
django-celery-results==2.5.1
django_celery_beat==2.6.0
flower==1.2.0
pika==1.3.2